Removed code from /hw into /swap and cleanup
